exmap: move action_params to public header
