+++ module/linux/exmap.h
@@ -12,7 +12,7 @@ struct exmap_ioctl_setup {
 };
 
 #define EXMAP_IOCTL_SETUP _IOC(_IOC_WRITE, 'k', 1, sizeof(struct exmap_ioctl_setup))
-
+#define EXMAP_IOCTL_ACTION _IOC(_IOC_WRITE, 'k', 2, sizeof(struct exmap_action_params))
 #define EXMAP_IOCTL_CLONE _IOC(_IOC_WRITE, 'k', 3, sizeof(int))
 
 // Maximum Range of exmap_page.len
@@ -79,3 +79,10 @@ typedef enum exmap_flags exmap_flags;
 
 
 #define SYS_EXMAP_ACTION 400
+
+struct exmap_action_params {
+	uint16_t interface;
+	uint16_t iov_len;
+	uint16_t opcode; // exmap_opcode
+	uint64_t flags;  // exmap_flags
+};
